Made code modular
